requirements.txt
alembic==1.13.1
Authlib==1.2.0
fastapi==0.103.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
httpx==0.24.1
ipython==8.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
asyncpg==0.28.0
requests==2.31.0
SQLAlchemy==2.0.29
uvicorn==0.22.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy-utils==0.41.1
python-multipart==0.0.9
pyjwt==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pydantic-settings==2.0.2
pydantic[email]
fastapi-cache2[redis]
